Define styling and layout of the search bar container of the header.
Original attempt was to keep the input type as:
search

But it turned into a headache to remove the
default "x" icon when focused.

Hence the switch to type text.

While reluctant to use padding for positioning,
I've found that this is the most reliable method
to accomplish a positioning that imitates the
original version with millimetric precision.