Define layout and styling of announcements card
Styling this card was rather challenging because
of the limitations in font-size.

In the end, to make a font-size that looked decent
I allowed for it's elements to overflow, then hide
that and use transform to scale it down.

I doubt this would be pragmatic in a real world
setting but I learned something new and gained a
new perspective on lay outs.